refactor: region 패키지 내의 repository 객체 패키지 분리
