+++ android/app/src/main/java/com/on/turip/di/RepositoryModule.kt
@@ -2,12 +2,12 @@ package com.on.turip.di
 
 import com.on.turip.data.content.repository.DefaultContentRepository
 import com.on.turip.data.creator.repository.DefaultCreatorRepository
-import com.on.turip.data.region.DefaultRegionRepository
+import com.on.turip.data.region.repository.DefaultRegionRepository
 import com.on.turip.data.trip.repository.DefaultTripRepository
 import com.on.turip.data.userStorage.repository.DefaultUserStorageRepository
 import com.on.turip.domain.content.repository.ContentRepository
 import com.on.turip.domain.creator.repository.CreatorRepository
-import com.on.turip.domain.region.RegionRepository
+import com.on.turip.domain.region.repository.RegionRepository
 import com.on.turip.domain.settingsStorage.repository.UserStorageRepository
 import com.on.turip.domain.trip.repository.TripRepository
