+++ android/app/src/main/java/com/on/turip/domain/region/repository/RegionRepository.kt
@@ -1,4 +1,6 @@
-package com.on.turip.domain.region
+package com.on.turip.domain.region.repository
+
+import com.on.turip.domain.region.RegionCategory
 
 interface RegionRepository {
     suspend fun loadRegionCategories(isKorea: Boolean): Result<List<RegionCategory>>
